refactor: API
This change is implementing the api project as an actual api, which is split from every server software. Its using packetevents api as a base to still handle the gui clicking etc. but decoupled from paper and velocity.

The bukkit module is now handling the menu clicking on the bukkit side.

Like this a velocity module could be implemented easily.